% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createServer.R
\name{createServer}
\alias{createServer}
\title{Create a new EpivizServer object}
\usage{
createServer(port = 7123L, try_ports = FALSE, daemonized = NULL,
  verbose = FALSE)
}
\arguments{
\item{port}{Port to which server will listen to <int>}

\item{try_ports}{Try various ports until an open port is found <logical>}

\item{daemonized}{Run in background using httpuv's daemonized libuv server <logical>}

\item{verbose}{Print verbose output <logical>}
}
\value{
an 'EpivizServer' object
}
\description{
Create a new EpivizServer object
}

