% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedArray-class.R
\docType{class}
\name{IndexedArray-class}
\alias{IndexedArray}
\alias{IndexedArray-class}
\title{Class providing an indexed array (hashtable)}
\description{
Class providing an indexed array (hashtable)
}
\section{Methods}{

\describe{
\item{\code{append(item)}}{Append item to tail of array, returns id of item <int>}

\item{\code{empty()}}{Remove all items from array}

\item{\code{get(id)}}{Get item with given id<int>, returns <ANY>, returns NULL if no item with given id}

\item{\code{length()}}{Return number of items on array <int>}
}}

