% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedArray-class.R
\docType{class}
\name{IndexedArray}
\alias{IndexedArray}
\title{Class providing an indexed array (hashtable)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IndexedArray
}
\description{
Class providing an indexed array (hashtable)
}
\keyword{datasets}

