% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_writer}
\alias{json_writer}
\title{JSON writer used by this package}
\usage{
json_writer(x, method = "C")
}
\arguments{
\item{x}{object to write to json}

\item{method}{method used to write json}
}
\description{
Currently this just renames \code{\link{toJSON}} in the \code{rjson} package.
}
\seealso{
\code{\link{toJSON}}
}

