% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue-class.R
\docType{class}
\name{Queue-class}
\alias{Queue}
\alias{Queue-class}
\title{Class providing a queue data structure}
\description{
Class providing a queue data structure
}
\section{Methods}{

\describe{
\item{\code{empty()}}{Remove all items from queue}

\item{\code{has_more()}}{Return TRUE if there are more items in queue <logical>}

\item{\code{length()}}{Return the number of items in queue <int>}

\item{\code{pop()}}{Pop next item from queue (returns NULL if queue is empty)}

\item{\code{push(item)}}{Push <item> onto queue}
}}

