% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue-class.R
\docType{class}
\name{Queue}
\alias{Queue}
\title{Class providing a queue data structure}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Queue
}
\description{
Class providing a queue data structure
}
\keyword{datasets}

