% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createServer.R
\name{createServer}
\alias{createServer}
\title{Create a new EpivizServer object}
\usage{
createServer(
  port = 7123L,
  static_site_path = "",
  try_ports = FALSE,
  daemonized = NULL,
  verbose = FALSE,
  non_interactive = FALSE
)
}
\arguments{
\item{port}{(int) port to which server will listen to.}

\item{static_site_path}{(character) path to serve static html files.}

\item{try_ports}{(logical) try various ports until an open port is found.}

\item{daemonized}{(logical) run in background using httpuv's daemonized libuv server.}

\item{verbose}{(logical) print verbose output.}

\item{non_interactive}{(logical) run in non-interactive mode. For development purposes only.}
}
\value{
an \code{\link{EpivizServer}} object
}
\description{
Create a new EpivizServer object
}
\examples{
server <- createServer(port=7123,
                       verbose=TRUE
                       )
}
\seealso{
\code{\link{EpivizServer}} for the class of objects returned
}
