% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpivizServer-class.R
\docType{class}
\name{EpivizServer}
\alias{EpivizServer}
\title{Class providing WebSocket connection server}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EpivizServer
}
\description{
Class providing WebSocket connection server
}
\keyword{datasets}

